refactor(channel): consolidate composer input helpers
